implement getRecipes in remote data source

+++ app/src/main/java/br/com/ericksprengel/android/baking/data/Ingredient.java
@@ -13,7 +13,7 @@ public class Ingredient {
     public int id;
 
     @SerializedName("quantity")
-    public int quantity;
+    public float quantity;
 
     @SerializedName("measure")
     public String measure;
